LPS-129709 Move role constants to web module
